// dashboard/src/api/models/priceTier.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export interface PriceTier {
  /** @pattern ^-?\d{0,17}(?:\.\d{0,2})?$ */
  unit_amount: string;
  from_value: number;
  /** @nullable */
  to_value: number | null;
}